+++ admin/4000.ProductCategory/4030.ProductCategoryUpdate.php
@@ -21,32 +21,47 @@ if (!SYSAdmin() and !AppAdmin() and !Admin() and !Staff()) {
 }
 
 $ProductCategoryID                      = filter_input(INPUT_POST, 'ProductCategoryID');
-
 $ProductCategoryCatalogBeforeConvert    = filter_input(INPUT_POST, 'ProductCategoryCatalogBeforeConvert');
-if (file_exists("../assets/catalog/$ProductCategoryCatalogBeforeConvert")) {
-    unlink("../assets/catalog/$ProductCategoryCatalogBeforeConvert");
+
+if (isset($_FILES['ProductCategoryCatalog']) != null) {
+
+    if ($ProductCategoryCatalogBeforeConvert != null) {
+        if (file_exists("../assets/catalog/$ProductCategoryCatalogBeforeConvert")) {
+            unlink("../assets/catalog/$ProductCategoryCatalogBeforeConvert");
+        }
+    }
+
+    $ProductCategoryCatalog                 = $_FILES['ProductCategoryCatalog']['name'];
+    $Dir                                    = "../assets/catalog/";
+    $File                                   = $_FILES['ProductCategoryCatalog']['tmp_name'];
+    $ProductCategoryCatalogConvert          =  uniqid() . "-" . date('Y-m-d') . "-" . $ProductCategoryCatalog;
+    move_uploaded_file($File, $Dir . $ProductCategoryCatalogConvert);
+} else {
+    $ProductCategoryCatalogConvert          = $ProductCategoryCatalogBeforeConvert;
 }
 
 $ProductCategoryPhotoBeforeConvert    = filter_input(INPUT_POST, 'ProductCategoryPhotoBeforeConvert');
-if (file_exists("../assets/img/productcategoryphoto/$ProductCategoryPhotoBeforeConvert")) {
-    unlink("../assets/img/productcategoryphoto/$ProductCategoryPhotoBeforeConvert");
+
+
+if (isset($_FILES['ProductCategoryPhoto']) != null) {
+
+    if ($ProductCategoryPhotoBeforeConvert != null) {
+        if (file_exists("../assets/img/productcategoryphoto/$ProductCategoryPhotoBeforeConvert")) {
+            unlink("../assets/img/productcategoryphoto/$ProductCategoryPhotoBeforeConvert");
+        }
+    }
+
+    $ProductCategoryPhoto                   = $_FILES['ProductCategoryPhoto']['name'];
+    $Dir                                    = "../assets/img/productcategoryphoto/";
+    $File                                   = $_FILES['ProductCategoryPhoto']['tmp_name'];
+    $ProductCategoryPhotoConvert            = uniqid() . "-" . date('Y-m-d') . "-" . $ProductCategoryPhoto;
+    move_uploaded_file($File, $Dir . $ProductCategoryPhotoConvert);
+} else {
+    $ProductCategoryPhotoConvert            = $ProductCategoryPhotoBeforeConvert;
 }
 
 $StatusID                               = filter_input(INPUT_POST, 'StatusID');
 $ProductCategoryName                    = filter_input(INPUT_POST, 'ProductCategoryName');
-
-$ProductCategoryCatalog                 = $_FILES['ProductCategoryCatalog']['name'];
-$Dir                                    = "../assets/catalog/";
-$File                                   = $_FILES['ProductCategoryCatalog']['tmp_name'];
-$ProductCategoryCatalogConvert          =  uniqid() . "-" . date('Y-m-d') . "-" . $ProductCategoryCatalog;
-move_uploaded_file($File, $Dir . $ProductCategoryCatalogConvert);
-
-$ProductCategoryPhoto                   = $_FILES['ProductCategoryPhoto']['name'];
-$Dir                                    = "../assets/img/productcategoryphoto/";
-$File                                   = $_FILES['ProductCategoryPhoto']['tmp_name'];
-$ProductCategoryPhotoConvert            = uniqid() . "-" . date('Y-m-d') . "-" . $ProductCategoryPhoto;
-move_uploaded_file($File, $Dir . $ProductCategoryPhotoConvert);
-
 $UpdateBy                               = filter_input(INPUT_POST, 'UpdateBy');
 $EventLogUser                           = $UpdateBy;
 $EventLogData                           = 'Update Product Category ' . $ProductCategoryName;
